+++ wicketd-commission-types/versions/src/lib.rs
@@ -26,6 +26,7 @@
 //!
 //! For more information, see RFD 619.
 
+mod impls;
 pub mod latest;
 #[path = "initial/mod.rs"]
 pub mod v1;
